created API request and response for Service/Package, created ServiceItem and lazycolumn in AssignedServicesTab

+++ app/src/main/java/com/inovatech/smartpack/model/Deliveryman.kt
@@ -4,7 +4,7 @@ import com.inovatech.smartpack.model.api.DeliverymanRequest
 
 data class Deliveryman(
     val id: Long = 0,
-    val userId: Int = 0,
+    val userId: Long = 0,
     val licence: String = "",
     val active: Boolean = true,
     val vehicle: Vehicle = Vehicle()
