+++ app/src/main/java/com/inovatech/smartpack/model/api/DeliverymanRequest.kt
@@ -3,6 +3,6 @@ package com.inovatech.smartpack.model.api
 import com.google.gson.annotations.SerializedName
 
 data class DeliverymanRequest(
-    @SerializedName("usuariId") val userId: Int,
+    @SerializedName("usuariId") val userId: Long,
     @SerializedName("llicencia") val licence: String
 )
